requirements.txt
mysql-connector-python==8.0.26
pandas ==1.3.2
requests ==2.26.0
circuitbreaker ==1.3.1
APScheduler==3.3.1